added all tables, idx and triggers for products, billboard and blog. Created procedure for fullProduct data join and manage in dto/controller

+++ src/main/resources/db/migration/V3__create_product_categories_table.sql
@@ -1,5 +1,5 @@
 create table if not exists product_categories(
-id int auto_increment PRIMARY KEY NOT NULL,
+id BIGINT auto_increment PRIMARY KEY NOT NULL,
 name VARCHAR(35) NOT NULL,
 description VARCHAR(255)
 );
