+++ src/main/resources/db/migration/V4__create_product_tags_table.sql
@@ -1,5 +1,5 @@
 create table if not exists product_tags(
-id int auto_increment PRIMARY KEY NOT NULL,
+id BIGINT auto_increment PRIMARY KEY NOT NULL,
 name VARCHAR(50) NOT NULL,
 description VARCHAR(255),
 discount_percentage DECIMAL(5,2)
